feat(outlet): manage outlet
