+++ apps/api/src/controllers/storeController/index.ts
@@ -1,32 +1,9 @@
-import prisma from '@/connection';
-import axios from 'axios';
+import { createStoreByAdminService, getAllStoreService, getStoreService } from '@/service/storeService';
 import { Request, Response, NextFunction } from 'express'
 
-interface IStoreMap {
-    id: string
-    storeName: string
-    address: string
-    city: string
-    province: string
-    country: string
-    zipCode: string
-    latitude: number
-    longitude: number
-    createdAt: Date
-    updatedAt: Date
-    deletedAt: Date | null
-}
-
-const rajaOngkirApiKey: string | undefined = process.env.RAJAONGKIR_API_KEY as string
 export const getStore = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const findStore = await prisma.store.findMany()
-        const dataStore = findStore?.map((store: IStoreMap) => {
-            return {
-                storeId: store?.id,
-                storeName: store?.storeName
-            }
-        })
+        const { dataStore } = await getStoreService()
 
         res.status(201).json({
             error: false,
@@ -44,40 +21,12 @@ export const getAllStore = async (req: Request, res: Response, next: NextFunctio
 
         const take = parseInt(limit as string)
         const skip = (parseInt(page as string) - 1) * take
-        let whereClause;
 
-        if (search) {
-            whereClause = {
-                OR: [
-                    { storeName: { contains: search as string } },
-                    { city: { contains: search as string } },
-                    { province: { contains: search as string } },
-                ]
-            }
-        }
+        const limitTypes = typeof limit == 'string' ? limit : ''
+        const searchTypes = typeof search == 'string' ? search : ''
+        const sortTypes = typeof sort == 'string' ? sort : ''
 
-        let findStore
-        if (sort == 'name-asc') {
-            findStore = await prisma.store.findMany({
-                where: whereClause, take, skip, orderBy: { storeName: 'asc' }
-
-            })
-        } else if (sort == 'name-desc') {
-            findStore = await prisma.store.findMany({
-                where: whereClause, take, skip, orderBy: { storeName: 'desc' }
-            })
-        } else {
-            findStore = await prisma.store.findMany({
-                where: whereClause, take, skip
-            })
-        }
-
-        const totalData = await prisma.store.count({
-            where: whereClause
-        })
-
-        const totalPage = Math.ceil(totalData / Number(limit))
-        if (findStore?.length == 0) throw { msg: 'Data outlet belum tersedia', status: 404 }
+        const { totalPage, findStore } = await getAllStoreService({ limit: limitTypes, search: searchTypes, sort: sortTypes, skip, take })
 
         res.status(200).json({
             error: false,
@@ -92,36 +41,8 @@ export const getAllStore = async (req: Request, res: Response, next: NextFunctio
 export const createStoreByAdmin = async (req: Request, res: Response, next: NextFunction) => {
     try {
         const { storeName, address, city, province, zipCode, latitude, longitude } = req.body
-        const responseApi: any = await axios.get(`https://api.rajaongkir.com/starter/province?id=${province}`, {
-            headers: { key: rajaOngkirApiKey }
-        })
 
-        if (!responseApi) throw { msg: 'Gagal mendapatkan data provinsi', status: 400 }
-        const provinceName: string = responseApi?.data?.rajaongkir?.results?.province
-
-        const findExistingStore = await prisma.store.findFirst({
-            where: {
-                AND: [
-                    { storeName },
-                    { address },
-                    { province: provinceName },
-                ]
-            }
-        })
-
-        if (findExistingStore) throw { msg: 'Outlet sudah tersedia harap tambah tempat lain', status: 400 }
-        await prisma.store.create({
-            data: {
-                storeName,
-                address,
-                city,
-                province: provinceName,
-                country: 'Indonesia',
-                zipCode,
-                latitude: parseFloat(latitude),
-                longitude: parseFloat(longitude)
-            }
-        })
+        await createStoreByAdminService({ storeName, address, city, province, zipCode, latitude, longitude })
 
         res.status(201).json({
             error: false,
