+++ apps/api/src/router/workerRouter/index.ts
@@ -2,9 +2,9 @@ import { Router } from 'express'
 import { limiter } from '@/middleware/rateLimit'
 import { tokenValidation } from '@/middleware/verifyToken'
 import { expressValidatorErrorHandling } from '@/middleware/validation/errorHandlingValidator'
-import {  changePasswordWorker, deleteProfilePictureWorker, updateProfileWorker, getAllWorker, getSingleDataWorker, getSingleWorkerById, deleteDataWorkerById, getLaundryItems, deleteLaundryItems, updateLaundryItems, createLaundryItems } from '@/controllers/workerController'
+import { changePasswordWorker, deleteProfilePictureWorker, updateProfileWorker, getAllWorker, getSingleDataWorker, getSingleWorkerById, deleteDataWorkerById } from '@/controllers/workerController'
 import { roleCheckSuperAdmin } from '@/middleware/roleCheck'
-import { changePasswordWorkerValidation, productLaundryValidation, updateProfileWorkerValidation } from '@/middleware/validation'
+import { changePasswordWorkerValidation, updateProfileWorkerValidation } from '@/middleware/validation'
 import { uploader } from '@/middleware/uploader'
 
 const workerRouter = Router()
@@ -33,19 +33,13 @@ const workerRouter = Router()
 // workerRouter.get('/history-ironing/', tokenValidation, getIroningHistory)
 // workerRouter.get('/history-packing/', tokenValidation, getPackingHistory)
 
-/* core */
 workerRouter.get('/', tokenValidation, getSingleDataWorker)
 workerRouter.patch('/profile', tokenValidation, uploader, limiter, updateProfileWorkerValidation, expressValidatorErrorHandling, updateProfileWorker)
 workerRouter.patch('/change-password', tokenValidation, limiter, changePasswordWorkerValidation, expressValidatorErrorHandling, changePasswordWorker)
 workerRouter.delete('/', tokenValidation, deleteProfilePictureWorker)
 
-/* super admin */
 workerRouter.get('/all-workers', getAllWorker)
 workerRouter.get('/detail/:id', getSingleWorkerById)
 workerRouter.delete('/detail/:id', tokenValidation, roleCheckSuperAdmin, limiter, deleteDataWorkerById)
-workerRouter.get('/laundry-items', tokenValidation, roleCheckSuperAdmin, getLaundryItems)
-workerRouter.post('/laundry-items', tokenValidation, roleCheckSuperAdmin, limiter, productLaundryValidation, createLaundryItems)
-workerRouter.delete('/laundry-items/:id', tokenValidation, roleCheckSuperAdmin, limiter, deleteLaundryItems)
-workerRouter.patch('/laundry-items/:id', tokenValidation, roleCheckSuperAdmin, limiter, productLaundryValidation, updateLaundryItems)
 
 export default workerRouter
