+++ apps/api/src/service/contactService/index.ts
@@ -1,0 +1,43 @@
+import prisma from "@/connection"
+import { validateEmail } from "@/middleware/validation/emailValidation"
+import { phoneNumberValidation } from "@/middleware/validation/phoneNumberValidation"
+import { compile } from "handlebars"
+import fs from 'fs'
+import { transporter } from "@/utils/transporter"
+import { IContactMessage } from "./types"
+
+export const createContactMessageService = async ({ email, phoneNumber, userId, textHelp, name }: IContactMessage) => {
+    if (!validateEmail(email)) throw { msg: 'Harap masukan format email dengan benar', status: 401 }
+    if (!phoneNumberValidation(phoneNumber)) throw { msg: 'Harap masukan format dengan angka', status: 401 }
+    const messageUser = await prisma.contact.create({ data: { name, email, phoneNumber, userId: userId, textHelp } })
+
+    if (messageUser) {
+        const emailFile = fs.readFileSync('./src/public/sendMail/emailContact.html', 'utf-8')
+        let compiledHtml: any = compile(emailFile)
+        compiledHtml = compiledHtml({ firstName: name, url: 'http://localhost:3000' })
+        await transporter.sendMail({
+            to: email,
+            subject: 'Terimakasih telah menghubungi kami',
+            html: compiledHtml
+        })
+    }
+}
+
+export const getContactMessageService = async () => {
+    const findMessageUser = await prisma.contact.findMany({
+        where: { deletedAt: null },
+        include: {
+            User: {
+                select: {
+                    firstName: true,
+                    lastName: true,
+                    profilePicture: true,
+                    phoneNumber: true,
+                    email: true,
+                },
+            }
+        }
+    })
+
+    return { findMessageUser }
+}
