+++ apps/api/src/service/contactService/types.ts
@@ -1,0 +1,7 @@
+export interface IContactMessage {
+    email: string,
+    phoneNumber: string,
+    userId: string,
+    textHelp: string,
+    name: string
+}
