+++ apps/api/src/service/laundryItemService/types.ts
@@ -1,0 +1,6 @@
+export interface IGetLaundryItems {
+    page?: string;
+    limit?: string;
+    search?: string;
+    sort?: string;
+}
