+++ apps/web/src/app/(worker)/worker/admin-outlet/c/[slug]/page.tsx
@@ -98,7 +98,7 @@ export default function Page({ params }: { params: Promise<{ slug: string }> }) 
     const { data: dataItemName, isLoading: dataItemNameLoading } = useQuery({
         queryKey: ['get-data-item'],
         queryFn: async () => {
-            const res = await instance.get('/laundry/', {
+            const res = await instance.get('/laundry', {
                 headers: { Authorization: `Bearer ${token}` }
             });
             console.log(res, 'itemname');
