+++ apps/web/src/features/superAdmin/hooks/useProductLaundryHooks/index.ts
@@ -22,7 +22,7 @@ const useProductLaundryHooks = () => {
     const { data: dataItem, isFetching, refetch } = useQuery({
         queryKey: ['get-data-item', searchItem],
         queryFn: async () => {
-            const response = await instance.get('/worker/laundry-items', {
+            const response = await instance.get('/laundry/laundry-items', {
                 params: {
                     search: searchItem,
                     page: currentPage,
@@ -39,7 +39,7 @@ const useProductLaundryHooks = () => {
 
     const { mutate: createProductItem, isPending } = useMutation({
         mutationFn: async ({ itemName }: { itemName: string }) => {
-            return await instance.post('/worker/laundry-items', { itemName }, {
+            return await instance.post('/laundry', { itemName }, {
                 headers: {
                     Authorization: `Bearer ${token}`
                 }
@@ -65,7 +65,7 @@ const useProductLaundryHooks = () => {
 
     const { mutate: handleDeleteItem, isPending: isPendingDelete } = useMutation({
         mutationFn: async (id: number) => {
-            return await instance.delete(`/worker/laundry-items/${id}`, {
+            return await instance.delete(`/laundry/laundry-items/${id}`, {
                 headers: {
                     Authorization: `Bearer ${token}`
                 }
@@ -92,7 +92,7 @@ const useProductLaundryHooks = () => {
 
     const { mutate: handleUpdateItem, isPending: isPendingUpdate } = useMutation({
         mutationFn: async ({ id, itemName }: { id: string, itemName: string }) => {
-            return await instance.patch(`/worker/laundry-items/${id}`, { itemName }, {
+            return await instance.patch(`/laundry/laundry-items/${id}`, { itemName }, {
                 headers: {
                     Authorization: `Bearer ${token}`
                 }
